% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-austria_boundary.R
\docType{data}
\name{austria_boundary}
\alias{austria_boundary}
\title{Austria Boundary Dataset}
\format{
A simple feature collection (sf) with 1 feature and 1 field:
\describe{
\item{FID}{Feature ID, a unique identifier for the feature.}
\item{geometry}{A POLYGON representing the boundary of Austria.}
}
}
\source{
\url{https://www.data.gv.at}
}
\usage{
austria_boundary
}
\description{
A simple feature collection representing the boundary of Austria.
}
\keyword{datasets}
