% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{poly4AT_processor}
\alias{poly4AT_processor}
\title{Open Shiny App}
\usage{
poly4AT_processor()
}
\value{
A Shiny application.
}
\description{
Launches the Poly4AT Shiny application for querying geographic data based on coordinates.
}
\examples{
\dontrun{
  poly4AT_processor()
}


}
